eve: change Scale and Translate param types

+++ eve/examples/gameduino/nightstrike/utils.go
@@ -42,6 +42,6 @@ func prob(num, denom int) bool {
 	return int(rand.Int63())&0xFFFF < num<<16/denom
 }
 
-func f16(n int) int32 {
-	return int32(n << 16)
+func f16(n int) int {
+	return n << 16
 }
